Use surface_transform in replay_and_create_regions
The surface_transform was already used surface_replay_with_clip, as the
matrix is obviously needed for the clip. But now, because of the
optimization done in commit 09b42c7, it's also needed by
replay_and_create_regions: get_target_extents clips the target surface
for performance issues, and therefore needs the surface_transform matrix
to get the right clipping surface.

Signed-off-by: Guillaume Ayoub <guillaume.ayoub@kozea.fr>