Un-doxygen disabled cairo_set_opacity
The cairo_set_opacity() routine has been disabled since 2010.  There
appears to be support for it internally in the backend, so it is unclear
why it is not enabled or what might break if it were re-enabled.  Given
the lack of bug reports about its omission, it may be vestigal.  For
now, just hide the doxygen for it to suppress warnings.  It may be worth
deeper review to either re-enable it or remove it entirely.

Reported-by: Rafał Mużyło (See fdo bugzilla #82741)
Signed-off-by: Bryce Harrington <bryce@osg.samsung.com>