cogl: Support CAIRO_EXTEND_NONE correctly
Prior to this change, CAIRO_EXTEND_NONE was handled the same way as
CAIRO_EXTEND_PAD, which produced incorrect results in many tests.
This fixes such behaviour by created a per-source and per-mask clip
that is applied so that areas outside the bounds of the source or
mask textures are not sampled during drawing.

Signed-off-by: George Matsumura <gmmatsumura01@bvsd.org>