cogl: Use less memory during recording surface replaying
Many of the tests in the test suite do many replays of different
small parts of a single surface. When combined with the cogl backend's
journaling mechanism, this consumes very large amounts of memory, as
the entire surface is allocated for each replay. This changes this
behaviour so that only the portion of the surface that will be used
is allocated.

Signed-off-by: George Matsumura <gmmatsumura01@bvsd.org>