cogl: Add minimal font support
This adds very basic path-based font support to the cogl backend,
improving cairo-perf-trace performance by 6x or more on many traces.
This still is not the most desirable form of font support, which
would resemble the texture-based glyph caching in the gl backend or
use cogl-pango to automatically cache glyphs.

Signed-off-by: George Matsumura <gmmatsumura01@bvsd.org>